Move FileListFilter to new package filters
