+++ Kitodo/src/main/java/de/sub/goobi/export/download/ExportPdf.java
@@ -33,7 +33,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.commons.httpclient.methods.GetMethod;
-import org.goobi.io.FileListFilter;
+import org.kitodo.filters.FileListFilter;
 import org.kitodo.data.database.beans.Process;
 import org.kitodo.services.ServiceManager;
 import org.kitodo.services.file.FileService;
